+++ lib/src/config/is_ios.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+extension BuildContextIsIOS on BuildContext {
+  bool get isIOS => Theme.of(this).platform == TargetPlatform.iOS;
+}