+++ lib/src/widgets/_.dart
@@ -1,5 +1,0 @@
-import 'dart:io';
-
-import 'package:flutter/foundation.dart';
-
-final bool isIOS = !kIsWeb && Platform.isIOS;